requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
httpx==0.28.1
